Changed the code according to Tim Juic suggestions.
