+++ app/src/main/java/com/Found404/paypro/ui/components/DeleteTerminalPopup.kt
@@ -1,0 +1,42 @@
+package com.Found404.paypro.ui.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun DeleteTerminalPopup(terminalId: String, onConfirm: () -> Unit, onCancel: () -> Unit) {
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(16.dp)
+            .background(color = Color.White, shape = RoundedCornerShape(16.dp))
+    ) {
+        Column(
+            modifier = Modifier.padding(16.dp),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Text("Please confirm that you want to delete terminal $terminalId.")
+            Spacer(modifier = Modifier.height(8.dp))
+            Button(onClick = onConfirm) {
+                Text("Confirm")
+            }
+            Spacer(modifier = Modifier.height(8.dp))
+            Button(onClick = onCancel) {
+                Text("Cancel")
+            }
+        }
+    }
+}
